requirements.txt
# Requirements for the CTF Multi-Agent System

# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Agent framework
openai>=1.0.0  # For AI-powered agents
anthropic>=0.8.0  # Alternative AI provider

# Communication and async
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
redis>=5.0.0  # Optional: for message queue

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.13.0
pymongo>=4.6.0  # Optional: for MongoDB

# Web scraping and requests
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.16.0
playwright>=1.40.0

# Cryptography
cryptography>=41.0.0
pycryptodome>=3.19.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Logging and monitoring
structlog>=23.2.0
sentry-sdk>=1.39.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
click>=8.1.0  # CLI framework
rich>=13.7.0  # Terminal formatting
python-magic>=0.4.27  # File type detection

# CTF-specific tools (Python libraries)
pwntools>=4.11.0
z3-solver>=4.12.0
capstone>=5.0.0  # Disassembly
keystone-engine>=0.9.2  # Assembly
unicorn>=2.0.0  # CPU emulation

# Optional: ML/AI
# scikit-learn>=1.3.0
# tensorflow>=2.15.0
# torch>=2.1.0
scapy
